Variable names and performance tweaking
Changed variable names from keywords in multiple places.
Removed density as a controlling factor and replaced by gridSize. Changed countGridpoints to work with gridSize instead of density.  Simplified a lot of the logic and scaling of the small Cluster grid size to be same as 1. changed the threshold in sparsity annotate.